+++ components/index.js
@@ -35,6 +35,7 @@ export * from './tbt-playground.js';
 export * from './tbt-icon.js';
 export * from './tbt-badge.js';
 export * from './tbt-chip.js';
+export * from './tbt-tooltip.js';
 export * from './tbt-alert.js';
 export * from './tbt-toast.js';